change module and file names to match version of earliest supported version of spring-context
