--- dd-java-agent/instrumentation/spring-scheduling-1/spring-scheduling-1.gradle
@@ -1,15 +1,0 @@
-muzzle {
-  pass {
-    group = 'org.springframework'
-    module = 'spring-context'
-    versions = "[5.2.3.RELEASE,]"
-    assertInverse = true
-  }
-}
-
-apply from: "${rootDir}/gradle/java.gradle"
-dependencies {
-  compileOnly group: 'org.springframework', name: 'spring-context', version: '5.2.3.RELEASE'
-  testCompile group: 'org.springframework', name: 'spring-context', version: '5.2.3.RELEASE'
-  testCompile project(':dd-java-agent:instrumentation:trace-annotation')
-}
