--- dd-java-agent/instrumentation/spring-scheduling-1/src/test/java/IntervalTask.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2012-2015 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Component;
-
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-
-@Component
-public class IntervalTask implements Runnable {
-
-  private final CountDownLatch latch = new CountDownLatch(1);
-
-  @Scheduled(initialDelay = 2, fixedRate = 5000)
-  @Override
-  public void run() {
-    latch.countDown();
-  }
-
-  public void blockUntilExecute() throws InterruptedException {
-    latch.await(5, TimeUnit.SECONDS);
-  }
-}
