+++ dd-java-agent/instrumentation/spring-scheduling-3.1/src/test/groovy/SpringSchedulingTest.groovy
@@ -4,44 +4,15 @@ import org.springframework.context.annotation.AnnotationConfigApplicationContext
 
 class SpringSchedulingTest extends AgentTestRunner {
 
-  def "schedule interval test"() {
-    setup:
-    def context = new AnnotationConfigApplicationContext(IntervalTaskConfig)
-    def task = context.getBean(IntervalTask)
-
-    TEST_WRITER.clear()
-
-    task.blockUntilExecute();
-
-    expect:
-    assert task != null;
-    assertTraces(1) {
-      trace(0, 1) {
-        span(0) {
-          resourceName "IntervalTask.run"
-          operationName "scheduled.call"
-          parent()
-          errored false
-          tags {
-            "$Tags.COMPONENT" "spring-scheduling"
-            defaultTags()
-          }
-        }
-      }
-    }
-  }
-
   def "schedule trigger test according to cron expression"() {
     setup:
     def context = new AnnotationConfigApplicationContext(TriggerTaskConfig)
     def task = context.getBean(TriggerTask)
 
-    TEST_WRITER.clear()
-
-    task.blockUntilExecute();
+    task.blockUntilExecute()
 
     expect:
-    assert task != null;
+    assert task != null
     assertTraces(1) {
       trace(0, 1) {
         span(0) {
@@ -56,5 +27,31 @@ class SpringSchedulingTest extends AgentTestRunner {
         }
       }
     }
+  }
+  
+  def "schedule interval test"() {
+    setup:
+    def context = new AnnotationConfigApplicationContext(IntervalTaskConfig)
+    def task = context.getBean(IntervalTask)
+
+    task.blockUntilExecute()
+
+    expect:
+    assert task != null
+    assertTraces(1) {
+      trace(0, 1) {
+        span(0) {
+          resourceName "IntervalTask.run"
+          operationName "scheduled.call"
+          parent()
+          errored false
+          tags {
+            "$Tags.COMPONENT" "spring-scheduling"
+            defaultTags()
+          }
+        }
+      }
+    }
+
   }
 }
