+++ settings.gradle
@@ -130,7 +130,7 @@ include ':dd-java-agent:instrumentation:servlet:request-3'
 include ':dd-java-agent:instrumentation:slf4j-mdc'
 include ':dd-java-agent:instrumentation:sparkjava-2.3'
 include ':dd-java-agent:instrumentation:spring-data-1.8'
-include ':dd-java-agent:instrumentation:spring-scheduling-1'
+include ':dd-java-agent:instrumentation:spring-scheduling-3.1'
 include ':dd-java-agent:instrumentation:spring-webmvc-3.1'
 include ':dd-java-agent:instrumentation:spring-webflux-5'
 include ':dd-java-agent:instrumentation:spymemcached-2.12'
